Extract User.last_billing_cycle to mixin
